Add command line option to automatically load all spectra...
This is for the case where a file to load is given on the command line,
and that file contains multiple spectra. By default, a dialog will pop
up to allow the user to select which spectra should be loaded. However,
with this option, all spectra will be loaded automatically. This option
also helps enable automatic tests since it avoids creating a popup
dialog.